REF: move string arithmetic tests to tests.arithmetic.test_string
